seethos-bot-manager/requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
altair==5.1.1
annotated-types==0.5.0
anyio==3.7.1
asttokens==2.4.1
async-timeout==4.0.3
attrs==23.1.0
blinker==1.6.2
cachetools==5.3.1
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
colorama==0.4.6
dataclasses-json==0.6.1
decorator==5.1.1
dnspython==2.4.2
docker==6.1.3
edge-tts==6.1.9
elevenlabs==0.2.27
executing==2.0.1
faiss-cpu==1.7.4
fastapi==0.103.1
ffmpeg-python==0.2.0
frozenlist==1.4.0
future==1.0.0
gitdb==4.0.10
GitPython==3.1.37
greenlet==2.0.2
gTTS==2.3.2
h11==0.16.0
idna==3.4
importlib-metadata==6.8.0
ipython==8.24.0
jedi==0.19.1
Jinja2==3.1.2
jsonpatch==1.33
jsonpointer==2.4
jsonschema==4.19.1
jsonschema-specifications==2023.7.1
langchain==0.0.302
langsmith==0.0.41
markdown-it-py==3.0.0
MarkupSafe==2.1.3
marshmallow==3.20.1
matplotlib-inline==0.1.7
mdurl==0.1.2
mpmath==1.3.0
multidict==6.0.4
mypy-extensions==1.0.0
numexpr==2.10.0
numpy==1.26.4
openai==0.27.8
packaging==23.1
pandas==2.2.2
parso==0.8.4
pexpect==4.9.0
Pillow==9.5.0
prompt-toolkit==3.0.43
protobuf==4.24.3
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==13.0.0
pydantic==2.4.1
pydantic_core==2.10.1
pydeck==0.8.0
pydub==0.25.0
Pygments==2.16.1
pymongo==3.11.0
Pympler==1.0.1
PyPDF2==3.0.1
python-dateutil==2.8.2
python-multipart==0.0.6

pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0.1
referencing==0.35.1
regex==2024.4.28
requests==2.31.0
rich==13.7.1
rpds-py==0.18.0
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
SQLAlchemy==2.0.21
stack-data==0.6.3
starlette==0.27.0
streamlit==1.25.0
streamlit-chat==0.1.1
sympy==1.12
tabulate==0.9.0
tenacity==8.2.2
tiktoken==0.4.0
tokenizers==0.13.3
toml==0.10.2
toolz==0.12.1
tornado==6.4
tqdm==4.66.4
traitlets==5.14.0
typing-inspect==0.9.0
typing_extensions==4.11.0
tzdata==2024.1
tzlocal==4.3.1
urllib3==2.0.5
uvicorn==0.23.2
validators==0.28.1
watchdog==3.0.0
wcwidth==0.2.13
websocket-client==1.6.3
websockets==11.0.3
yarl==1.9.4
zipp==3.18.1
python-dotenv==0.19.1
